Move GUI startup code to a separate function
